backend/requirements-local.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-multipart==0.0.9
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
rich==13.8.1
websockets==12.0
